pyproject.toml
# SPDX-FileCopyrightText: The technology-data authors
#
# SPDX-License-Identifier: MIT

[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "technologydata"
dynamic = ["version"]
keywords = ["energy", "modelling", "techno-economics", "macroeconomics"] # TODO update
description = "Package for providing common data assumptions for energy system modelling on techno-economics and macro-economics."
readme = "README.md"
authors = [
    { name = "Contributors to technologydata", email = "johannes.hampp@openenergytransition.org" },
    { name = "Contributors to technologydata", email = "fabrizio.finozzi.business@gmail.com" },
]
classifiers=[
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
]
requires-python = ">=3.12"
dependencies = [
    "hdx-python-country>=3.9.6",
    "ipykernel>=6.29.5", # this one can go into optional as "dev"
    "mypy>=1.15.0",
    "pandas>=2.2.3",
    "pint>=0.24.4",
    "pre-commit>=4.2.0",
    "pydantic>=2.11.7",
    "pydeflate>=2.1.3", # this one can go into optional as "dev"
    "pytest>=8.3.5", # this one can go into optional as "dev"
    "requests>=2.32.3",
    "reuse>=5.0.2",
    "savepagenow>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/PyPSA/technology-data"
Source = "https://github.com/PyPSA/technology-data"
BugTracker = "https://github.com/PyPSA/technology-data/issues"
Changelog = "https://github.com/PyPSA/technology-data/releases"

[project.optional-dependencies]
docs = [
    "mkdocs-autolinks-plugin>=0.7.1",
    "mkdocs-git-revision-date-localized-plugin>=1.4.7",
    "mkdocs-material>=9.6.14",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocs-open-in-new-tab>=1.0.8",
    "mkdocstrings>=0.29.1",
    "mkdocstrings-python>=1.16.11",
]


# Setuptools_scm settings

[tool.uv]
package = true

[tool.setuptools_scm]
version_scheme = "no-guess-dev"

[tool.setuptools.packages.find]
include = ["technologydata"]

# Static type checker settings

[tool.mypy]
exclude = ['dev/*', 'examples/*', 'doc/*', 'test/*']
ignore_missing_imports = true
no_implicit_optional = true
warn_unused_ignores = true
show_error_code_links = true
disallow_any_generics = true
warn_return_any = true

[[tool.mypy.overrides]]
module = "technologydata"
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "requests"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "dateutil"
ignore_missing_imports = true

# Formatter and linter settings

[tool.ruff]
extend-include = ['*.ipynb']

[tool.ruff.lint]
select = [
    'F',   # pyflakes
    'E',   # pycodestyle: Error
    'W',   # pycodestyle: Warning
    'I',   # isort
    'D',   # pydocstyle
    'UP',  # pyupgrade
    'ANN', # flake-8 annotations
    'TID', # flake8-tidy-imports
    'NPY', # numpy
    'PD',  # pandas
    'RUF013', # ruff
]

ignore = [
    'ANN401', # Dynamically typed expressions are forbidden
    'E501', # line too long
    'D203', # 1 blank line required before class docstring
    'D212', # Multi-line docstring summary should start at the second line
    ]


[tool.ruff.lint.per-file-ignores]
"{test}/**"=[
    'ANN' # flake8-annotations
]

[tool.ruff.lint.flake8-tidy-imports]
# Disallow all relative imports.
ban-relative-imports = "all"
